+++ src/main/java/com/epam/ak/model/model/SemiPreciousGem.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public class SemiPreciousGem extends Gem {
     private Tracery tracery;
 
-    public SemiPreciousGem(UUID uuid, long id, String name, Type type, int transparency, Color color, int facet, Origin miningOrigin, int weight, int price, Tracery tracery) {
+    public SemiPreciousGem(UUID uuid, Integer id, String name, Type type, int transparency, Color color, int facet, Origin miningOrigin, int weight, int price, Tracery tracery) {
         super(uuid, id, name, type, transparency, color, facet, miningOrigin, weight, price);
         this.tracery = tracery;
     }
@@ -25,7 +25,7 @@ public class SemiPreciousGem extends Gem {
         public Tracery() {
         }
 
-        public Tracery(UUID uuid, long id, String name) {
+        public Tracery(UUID uuid, Integer id, String name) {
             super(uuid, id, name);
         }
     }
